Create WavePrintMatrixProgram.java
This is the program for the "Wave Print Matrix Program" addressed in the Issue Number 222. I believe this is exactly what was asked to do, so kindly merge it into the repo. This is the Java Version of the same, i will be making and requesting for the C++ version too, so kindly see that too in PRs.

Link to the issue : https://github.com/panditakshay402/Hacktoberfest2023/issues/222