Create WavePrintMatrixProgram.cpp
This is the program for the "Wave Print Matrix Program" addressed in the Issue Number 222. I believe this is exactly what was asked to do, so kindly merge it into the repo. This is the C++ Version of the same, i have also requested for the Java version too in my previous PR, so kindly see that too if it haven't been merged yet.

Link to the issue : https://github.com/panditakshay402/Hacktoberfest2023/issues/222